requirements.txt
datasets 
sentence-transformers 
faiss-cpu
numpy
scikit-learn
nltk
tqdm
joblib
torch